+++ example/lib/pages/mini_games/simple_example/my_enemy.dart
@@ -12,24 +12,31 @@ import 'package:example/shared/util/enemy_sprite_sheet.dart';
 ///
 /// Rafaelbarbosatec
 /// on 19/10/21
-class MyEnemy extends SimpleEnemy with BlockMovementCollision {
+class MyEnemy extends SimpleNpc
+    with BlockMovementCollision, AutomaticRandomMovement {
   MyEnemy(Vector2 position)
       : super(
           animation: EnemySpriteSheet.simpleDirectionAnimation,
           position: position,
           size: Vector2.all(32),
-          life: 100,
         );
+
+  @override
+  Future<void> onLoad() {
+    add(RectangleHitbox(size: size));
+    return super.onLoad();
+  }
 
   @override
   void update(double dt) {
     super.update(dt);
-    seeAndMoveToPlayer(
-      margin: -5,
-      closePlayer: (player) {
-        /// do anything when close to player
-      },
-      radiusVision: 64,
-    );
+    // seeAndMoveToPlayer(
+    //   margin: -5,
+    //   closePlayer: (player) {
+    //     /// do anything when close to player
+    //   },
+    //   radiusVision: 64,
+    // );
+    runRandomMovement(dt);
   }
 }