+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-# next
-- fix bug in `AutomaticRandomMovement`. (not stop movement in collision)
-- fix bug in `BlockMovementCollision`.(pushing other)
+# 3.1.2
+- fix bug `AutomaticRandomMovement`. (not stop movement in collision)
+- fix bug `BlockMovementCollision`.(pushing other)
+- fix bug `simpleAttackMeleeByDirection`.(pushing enemy to inside other collision)
 
 ## 3.1.1
 - Fix flip render problem in `SimpleDirectionAnimation`.